refactor(config): split manifest model helpers
