+++ crates/project-model/src/toml_manifest/mod.rs
@@ -1,0 +1,9 @@
+mod schema;
+mod spans;
+
+pub(crate) use schema::TomlManifestSchema;
+pub use schema::{TomlManifestDiagnostic, toml_manifest_diagnostics};
+pub use spans::{
+    TomlManifestField, TomlManifestPath, toml_manifest_field_at_offset, toml_manifest_fields,
+    toml_manifest_path_at_offset, toml_manifest_paths,
+};
