+++ crates/project-model/src/toml_workspace.rs
@@ -1,208 +1,8 @@
-use std::{ops::Range, sync::LazyLock};
-
 use anyhow::Context;
-use const_format::formatcp;
 use itertools::Itertools;
-use regex::Regex;
-use rustc_hash::FxHashSet;
-use serde::Deserialize;
-use smol_str::SmolStr;
-use toml_edit::{ImDocument, Value};
-use utils::paths::{AbsPathBuf, Utf8PathBuf, sort_and_remove_subfolders};
+use utils::paths::{AbsPathBuf, sort_and_remove_subfolders};
 
-use crate::macro_def::{MacroAtom, MacroDef};
-
-const IDENTIFIER_RE: &str = r"[a-zA-Z_][a-zA-Z0-9$_]*|\\\S* ";
-static IDENT_RE: LazyLock<Result<Regex, regex::Error>> =
-    LazyLock::new(|| Regex::new(formatcp!("^({IDENTIFIER_RE})$")));
-static KV_RE: LazyLock<Result<Regex, regex::Error>> =
-    LazyLock::new(|| Regex::new(formatcp!("^({IDENTIFIER_RE})=(.*)$")));
-
-#[derive(Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct TomlManifestSchema {
-    #[serde(default)]
-    pub top_modules: Vec<String>,
-    #[serde(deserialize_with = "de_macros", default)]
-    pub defines: MacroDef,
-    #[serde(default)]
-    pub sources: Option<Vec<Utf8PathBuf>>,
-    #[serde(default)]
-    pub include_dirs: Option<Vec<Utf8PathBuf>>,
-    #[serde(default)]
-    pub libraries: Vec<Utf8PathBuf>,
-    #[serde(default)]
-    pub exclude: Vec<Utf8PathBuf>,
-}
-
-fn de_macros<'de, D>(deserializer: D) -> Result<MacroDef, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let res = Vec::<SmolStr>::deserialize(deserializer)?;
-    let ident_re = IDENT_RE.as_ref().map_err(|err| {
-        serde::de::Error::custom(format!("invalid macro identifier regex: {err}"))
-    })?;
-    let kv_re = KV_RE
-        .as_ref()
-        .map_err(|err| serde::de::Error::custom(format!("invalid macro key-value regex: {err}")))?;
-    let macros = res
-        .into_iter()
-        .map(|macr: SmolStr| {
-            if ident_re.is_match(&macr) {
-                Ok(MacroAtom::Flag(macr))
-            } else if let Some(caps) = kv_re.captures(&macr) {
-                let Some(key_match) = caps.get(1) else {
-                    return Err(serde::de::Error::custom(format!(
-                        "Invalid macro definition: {macr}"
-                    )));
-                };
-                let Some(value_match) = caps.get(2) else {
-                    return Err(serde::de::Error::custom(format!(
-                        "Invalid macro definition: {macr}"
-                    )));
-                };
-                let mut key: SmolStr = key_match.as_str().into();
-                let value = value_match.as_str().into();
-                if key.starts_with('\\') {
-                    let Some(stripped) =
-                        key.strip_prefix('\\').and_then(|key| key.strip_suffix(' '))
-                    else {
-                        return Err(serde::de::Error::custom(format!(
-                            "Invalid escaped macro name: {macr}"
-                        )));
-                    };
-                    key = stripped.into();
-                }
-                Ok(MacroAtom::KeyValue { key, value })
-            } else {
-                Err(serde::de::Error::custom(format!("Invalid macro definition: {macr}")))
-            }
-        })
-        .collect::<Result<Vec<_>, _>>()?
-        .into_iter()
-        .collect::<FxHashSet<_>>();
-    Ok(MacroDef { macros })
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct TomlManifestDiagnostic {
-    pub range: Option<Range<usize>>,
-    pub message: String,
-}
-
-pub fn toml_manifest_diagnostics(text: &str) -> Vec<TomlManifestDiagnostic> {
-    match toml::from_str::<TomlManifestSchema>(text) {
-        Ok(_) => Vec::new(),
-        Err(error) => {
-            vec![TomlManifestDiagnostic { range: error.span(), message: error.to_string() }]
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct TomlManifestField {
-    pub key: String,
-    pub key_range: Range<usize>,
-    pub value_range: Range<usize>,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct TomlManifestPath {
-    pub key: String,
-    pub value: String,
-    pub value_range: Range<usize>,
-    pub content_range: Range<usize>,
-}
-
-pub fn toml_manifest_fields(text: &str) -> Vec<TomlManifestField> {
-    manifest_top_level_values(text)
-        .into_iter()
-        .filter_map(|(key, key_range, value)| {
-            Some(TomlManifestField { key, key_range, value_range: value.span()? })
-        })
-        .collect()
-}
-
-pub fn toml_manifest_field_at_offset(text: &str, offset: usize) -> Option<TomlManifestField> {
-    toml_manifest_fields(text)
-        .into_iter()
-        .find(|field| range_contains_offset(&field.key_range, offset))
-}
-
-pub fn toml_manifest_paths(text: &str) -> Vec<TomlManifestPath> {
-    manifest_top_level_values(text)
-        .into_iter()
-        .filter(|(key, _, _)| MANIFEST_PATH_FIELDS.contains(&key.as_str()))
-        .flat_map(|(key, _, value)| {
-            let mut paths = Vec::new();
-            if let Some(path) = manifest_string_value(&key, &value, text) {
-                paths.push(path);
-            }
-            if let Some(array) = value.as_array() {
-                paths.extend(
-                    array.iter().filter_map(|value| manifest_string_value(&key, value, text)),
-                );
-            }
-            paths
-        })
-        .collect()
-}
-
-pub fn toml_manifest_path_at_offset(text: &str, offset: usize) -> Option<TomlManifestPath> {
-    toml_manifest_paths(text)
-        .into_iter()
-        .find(|path| range_contains_offset(&path.content_range, offset))
-}
-
-fn manifest_top_level_values(text: &str) -> Vec<(String, Range<usize>, Value)> {
-    let Ok(document) = text.parse::<ImDocument<String>>() else {
-        return Vec::new();
-    };
-
-    document
-        .as_table()
-        .get_values()
-        .into_iter()
-        .filter_map(|(keys, value)| {
-            let [key] = keys.as_slice() else {
-                return None;
-            };
-            Some((key.get().to_string(), key.span()?, value.clone()))
-        })
-        .collect()
-}
-
-fn manifest_string_value(key: &str, value: &Value, text: &str) -> Option<TomlManifestPath> {
-    let value_range = value.span()?;
-    let content_range = string_content_range(text, value_range.clone())?;
-
-    Some(TomlManifestPath {
-        key: key.to_string(),
-        value: value.as_str()?.to_string(),
-        value_range,
-        content_range,
-    })
-}
-
-fn string_content_range(text: &str, value_range: Range<usize>) -> Option<Range<usize>> {
-    let raw = text.get(value_range.clone())?;
-    let quote_len = if raw.starts_with("\"\"\"") || raw.starts_with("'''") {
-        3
-    } else if raw.starts_with('"') || raw.starts_with('\'') {
-        1
-    } else {
-        return None;
-    };
-    (raw.len() >= quote_len * 2)
-        .then_some(value_range.start + quote_len..value_range.end - quote_len)
-}
-
-fn range_contains_offset(range: &Range<usize>, offset: usize) -> bool {
-    range.start <= offset && offset <= range.end
-}
-
-const MANIFEST_PATH_FIELDS: &[&str] = &["sources", "include_dirs", "libraries", "exclude"];
+use crate::{macro_def::MacroDef, toml_manifest::TomlManifestSchema};
 
 #[derive(Debug, PartialEq, Eq)]
 pub struct TomlWorkspace {
@@ -315,43 +115,6 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_de_macros() {
-        let toml = r#"
-top_modules = ["main"]
-defines = [
-    "foo",
-    "bar",
-    "FOO=bar",
-    "BAR=foo",
-    "BAZ=foo bar",
-    "eqwe=123",
-]
-        "#;
-        let toml_schema: TomlManifestSchema = toml::from_str(toml).unwrap();
-        assert_eq!(toml_schema.top_modules, ["main"]);
-        let mut macros = FxHashSet::default();
-        macros.insert(MacroAtom::Flag("foo".into()));
-        macros.insert(MacroAtom::Flag("bar".into()));
-        macros.insert(MacroAtom::KeyValue { key: "FOO".into(), value: "bar".into() });
-        macros.insert(MacroAtom::KeyValue { key: "BAR".into(), value: "foo".into() });
-        macros.insert(MacroAtom::KeyValue { key: "BAZ".into(), value: "foo bar".into() });
-        macros.insert(MacroAtom::KeyValue { key: "eqwe".into(), value: "123".into() });
-        assert_eq!(toml_schema.defines, MacroDef { macros });
-    }
-
-    #[test]
-    fn macro_predefines_are_stable() {
-        let toml = r#"
-defines = [
-    "BAR=foo",
-    "FOO",
-]
-        "#;
-        let toml_schema: TomlManifestSchema = toml::from_str(toml).unwrap();
-        assert_eq!(toml_schema.defines.to_predefine_strings(), ["BAR=foo", "FOO"]);
-    }
-
-    #[test]
     fn empty_manifest_uses_syntax_only_default() {
         let root = TestDir::new("empty-manifest");
         let manifest = root.write("vizsla_config.toml", "");
@@ -405,53 +168,5 @@ defines = [
         let workspace = TomlWorkspace::load_from_file(&manifest, false).unwrap();
 
         assert!(workspace.include_dirs.is_empty());
-    }
-
-    #[test]
-    fn manifest_diagnostics_report_schema_errors() {
-        let diagnostics = toml_manifest_diagnostics("source = [\"rtl\"]\n");
-
-        assert_eq!(diagnostics.len(), 1);
-        assert!(diagnostics[0].message.contains("unknown field"));
-    }
-
-    #[test]
-    fn manifest_field_lookup_uses_toml_key_spans() {
-        let toml = "sources = [\"rtl\"]\n";
-        let field = toml_manifest_field_at_offset(toml, 1).unwrap();
-
-        assert_eq!(field.key, "sources");
-        assert_eq!(field.key_range, 0..7);
-        assert_eq!(&toml[field.value_range], "[\"rtl\"]");
-        assert!(toml_manifest_field_at_offset(toml, 12).is_none());
-    }
-
-    #[test]
-    fn manifest_path_lookup_uses_toml_value_spans() {
-        let toml = "sources = [\n  \"rtl/top.sv\",\n]\n";
-        let offset = toml.find("top").unwrap();
-        let path = toml_manifest_path_at_offset(toml, offset).unwrap();
-
-        assert_eq!(path.key, "sources");
-        assert_eq!(path.value, "rtl/top.sv");
-        assert_eq!(&toml[path.content_range.clone()], "rtl/top.sv");
-        assert!(toml_manifest_path_at_offset(toml, 1).is_none());
-    }
-
-    #[test]
-    fn manifest_paths_lists_top_level_path_arrays() {
-        let toml = "sources = [\"rtl\", \"ip\"]\ntop_modules = [\"top\"]\n";
-        let paths = toml_manifest_paths(toml);
-        let values = paths.iter().map(|path| path.value.as_str()).collect::<Vec<_>>();
-
-        assert_eq!(values, ["rtl", "ip"]);
-    }
-
-    #[test]
-    fn manifest_path_lookup_ignores_non_path_fields() {
-        let toml = "top_modules = [\"top\"]\n";
-        let offset = toml.find("top").unwrap();
-
-        assert!(toml_manifest_path_at_offset(toml, offset).is_none());
     }
 }
